[refactor][#1] User 저장하는 패키지 이름 변경
