+++ src/main/java/kr/startoff/backend/repository/UserRepository.java
@@ -1,8 +1,7 @@
 package kr.startoff.backend.repository;
 
-import kr.startoff.backend.model.User;
+import kr.startoff.backend.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
