+++ src/main/java/kr/startoff/backend/security/jwt/JwtUtils.java
@@ -3,7 +3,7 @@ package kr.startoff.backend.security.jwt;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
-import kr.startoff.backend.model.User;
+import kr.startoff.backend.entity.User;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
